refactor: keep imports from utils minimal
